minor chanegs

+++ src/main/java/Obstacle.java
@@ -48,7 +48,7 @@ public class Obstacle {
 
             //g.fillRect(obstacle.getObstacleX(), obstacle.getObstacleY(),
                     //obstacle.getOBSTACLE_WIDTH(), obstacle.getObstacleHeight()); // Draws Obstacle hitbox for testing.
-            g.drawImage(obstacle.img, obstacle.getObstacleX(), obstacle.getObstacleY(),
+            g.drawImage(obstacle.img, (int) obstacle.getObstacleX(), obstacle.getObstacleY(),
                     obstacle.getOBSTACLE_WIDTH(), obstacle.getObstacleHeight(), null);
         }
     }
